Adds colorful sorted bar chart examples.
Specifically commit adds two examples:
1. Sorted bar chart featuring brewer color palette

This example depends on [palettable](https://jiffyclub.github.io/palettable/)
for the brewer color palettes.

It also shows how to sort on a specific field.

2. Pandas Series to sorted bar chart

This example also uses brewer colors as well as pandas and numpy directly.
It shows how to work with a Pandas Series object to create a DataFrame and
explicitly name the index and data points using a 'title' column.