Add support for generating vega spec in headless
This closes #441 by adding support for generating Vega spec's from vega
lite specs. It does this by using a headless browser (either chrome
or firefox) and loading a hosted version of the vega lite JS API.

To test this code install seleneium (`pip install selenium`) and
chromedriver (`brew cask install chromedriver`).

I implemented this with minimal changes to the existing API by adding
the `vega` format to the `spec_to_image_mimebundle` function. So it will
return a dictionary mapping the correct vega mimetype (as defined in
https://altair-viz.github.io/user_guide/display_frontends.html#renderer-api)
to the vega spec. It doesn't really make much sense to render vega to vega,
but this is currently supported as a no-op.