example: update trellis_area to use facet
See https://github.com/altair-viz/altair/issues/3178

This removes the old trellis_area example, because that is just a
simpler version of the sorted version.
Might as well just have the sorted version,
it is easy enough to see how you would jusst remove the sort
to get back to the original example.

This also adds some commas to the code so that
black keeps the formatting the same.

I wanted to update the name of the example so that it didn't contain
the word "trellis". While
I was at it I also wanted the examples to be in better sort order.