doc: add example, bidmensional 3d view
The chart created in this code is a two-dimensional visualization tool designed to represent three dimensions of data in an intuitive and interactive manner. It uses vehicle data, combining adjusted horsepower (categorized by color), adjusted weight (on the X-axis), and cylinder categories (distinguished in separate charts). This approach allows users to clearly and simply visualize the relationship between these variables, despite the inherent complexity of the data.

Moreover, the chart employs bubble size to represent the count of vehicles in each group, adding an additional layer of data to the analysis. Interactive features, such as selections and filters, enable users to customize the visualization, focusing on specific ranges of horsepower or other characteristics. This enhances the exploration and understanding of the underlying patterns and trends in the vehicle data.