refactor: Remove `@curried.curry` dependency from `limit_rows`
This pattern is what I've used in most cases. It works and type checks correctly, but could be improved as a decorator.
Would have the benefit of preserving the return type, if one were to inspect the intermediate function.

Also replacing the `TypeVar` with a `Union` seems to have satisfied `mypy`.