test: Change `skipif` -> `xfail` for `test_sanitize_pyarrow_table_columns` on `win32`
Seems to be a better fit, considering this would fail for `win32` devs not using `pyarrow` but have it installed.
This could be removed in the future if installing/confirming install `tzdata` to the correct location was handled during environment activation.

Possibly resolves: https://github.com/vega/altair/pull/3431#discussion_r1631904170