feat(typing): adds `Map` alias for `Mapping[str, Any]`
This allows a `TypedDict` to be used, where `mypy` previously would have required exactly a `dict`.

The name is not used anywhere currently in `altair` and has the benefit of being short like `dict`, but more permissive.

I have only added this to `Chart.encode` as I know this does not require the mutability of `dict`, which I cannot confidently say for elsewhere in `altair`.

Fixes https://github.com/vega/altair/pull/3427#discussion_r1662542866