docs: Use more consistent comments in *Failures on specific python version(s)*
These commands are referring to `hatch` environments named like `hatch-test.py3.8`, but not `doc` or `default`.
It is simpler to just describe this as python versions.

Also the [hatch test](https://hatch.pypa.io/latest/cli/reference/#hatch-test) docs refer to `--python` as accepting multiple versions, but do not give an example with more than one. I've not figured out how this would work, so this keeps it simple.