chore: Run `typos` and `codespell` against the codebase
These two tools found a number of spelling errors,
but their results had to be manually reviewed and,
in some cases, corrected.

For example, the word `ser` was corrected to `set`,
which was a change that had to be undone.

For example, `clearner` required a human guess
whether "clearer" or "cleaner" was intended.